Move utils into parser module
