+++ src/parser/mod.rs
@@ -1,3 +1,4 @@
+pub mod util;
 
 
 use SassItem;
@@ -11,7 +12,7 @@ use functions::SassFunction;
 use nom::IResult;
 #[cfg(test)]
 use num_rational::Rational;
-use parseutil::{comment, ignore_space, name, opt_spacelike, spacelike};
+use parser::util::{comment, ignore_space, name, opt_spacelike, spacelike};
 use selectors::selectors;
 use std::fs::File;
 use std::io::Read;
